Rewrite tbmconv to handle multiple files per archive
Rewrite tbmconv to handle archives containing multiple files. The behavior
of the program has been changed such that the output file name is treated
as a prefix (which gets the number of the file appended to it) if there
are multiple files in the archive; if there is only one file in the
archive, no numbers are appended. The program also now supports placing
the string "%d" somewhere in the file name to control where the file
number is written.

- Refactor code to read various structures (SYSLBN, VOL1, HDR1, HDR2)
  into functions.
- Add a new function, tbm_read(), which reads the HDR1/HDR2/EOF1 blocks
  surrounding each file and records the offset to the start of each file
  as well as the total file size (minus data buffer flags sprinkled
  throughout).
- Add two new structures to represent TBM files: TBMArchive and TBMFile.

Minor changes:
- Wrap long lines.
- Use 'return' instead of 'exit' in several locations.
- Improve error messages.